Rename generated JSON to clarify purpose
The clients access the protobuf definitions in `protos/` using
a generated JSON file previously kept at `src/protos.json`.
This has caused confusion about the purpose of the file, so it
has been renamed `generated_protos.json`, and a comment has been
added above the line that imports it.
