+++ fish_client/scripts/generate_proto_json.js
